Setup basic linting with ESLint and CodeClimate
This commit sets the recommended rules for ESLint and sets up the
CodeClimate configuration file. The default parser was changed to
Babel due to ESLint erroring with 'unexpected token' when using
the default parser provided. ESLint isn't configured yet to
reject/revise upon commit locally - the checks only run when
pushing to remote.