Add script to run ESLint on all files before commit
This commit adds a git hook to run ESLint on all staged files
(modified, added, or copied) and report which checks failed and
which checks passed. It lets the user commit anyway if they choose
in case ESLint is being overly picky. The script is modified
from this link:

https://gist.github.com/linhmtran168/2286aeafe747e78f53bf

Git hooks are not tracked by git and are local in the .git/hooks
folder. To get around this I created a .githooks directory and
moved the pre-commit hook into that folder. Each person must
run the command git config --local core-hooksPath .githooks/
in order to change the default git hooks directory to this new
.githooks directory under version control. If anybody has another
way to get around this that would be more ideal but this is the
simplest solution I could find in the moment.