Add pipeline docs in root directory
This commit moves the info from the pipeline status doc into an
.md file for the devs to read. It's meant to give instructions
on how to use the various parts of the pipeline. I updated
documentation for how to fix auto-fixable problems and also
about using AirBnB rules. We need to add how to run jest locally
and more info about the other parts of the pipeline.