Add JSDocs action
This commit adds a Github Action to run JSDocs on the source and
pipeline directories and output the html files in the docs
directory. Smaller implementation details of the changes are as
follows:

- I had to install jsdocs again (npm install --save-dev
jsdoc) because there was no jsdoc directory in node_modules/.bin.

- There's a configuration file that specifies what directories to
run on, and what the output directory should be. I specified this
to be 'docs'.

- In package.json I added a 'generate-docs' command for running
jsdocs with this config file in shorthand.

- The eslint pre-commit hook started running on generated .js
files in the docs directory, so I excluded that directory when
running the grep command.