Change script in package.json to fix docs action
This commit changes the command for generate-docs in package.json
as the command jsdoc is not recognized locally for some reason.

It also updates .eslintignore to ignore the docs directory, and
changes the configuration file for jsdocs back to its default
since the source and destination directories are now specified
in the command for generate-docs.