Pipeline phase 1.1 jsdocs (#19)
* Add JSDocs action

This commit adds a Github Action to run JSDocs on the source and
pipeline directories and output the html files in the docs
directory. Smaller implementation details of the changes are as
follows:

- I had to install jsdocs again (npm install --save-dev
jsdoc) because there was no jsdoc directory in node_modules/.bin.

- There's a configuration file that specifies what directories to
run on, and what the output directory should be. I specified this
to be 'docs'.

- In package.json I added a 'generate-docs' command for running
jsdocs with this config file in shorthand.

- The eslint pre-commit hook started running on generated .js
files in the docs directory, so I excluded that directory when
running the grep command.

* Test JSDocs Action by adding a Car class

This commit also fixes some indentation issues in the pipeline
directory that aren't in the linting example .js file.
Additionally it fixes the name of the job.

* Change script in package.json to fix docs action

This commit changes the command for generate-docs in package.json
as the command jsdoc is not recognized locally for some reason.

It also updates .eslintignore to ignore the docs directory, and
changes the configuration file for jsdocs back to its default
since the source and destination directories are now specified
in the command for generate-docs.

* Change docs action to specify output directory

* Change docs action to marketplace action

This commit is testing to see if changing the action I wrote
to the marketplace action works. Right now I'm thinking that
the previous action I wrote does not commit the generated html
files in the docs directory, which we need to do.

* Do not use docs-action directory to run action and test

* Add branch ref to test doc action

* Test adding deploy step in docs action

* Add deploy step and deploy_key

* Remove passphrase and test action

* Change publish branch to gh-pages

* Change ref to pipeline-phase-1.1-jsdocs

* Remove output directory from JSDocs action

* Add output directory back in jsdocs action

* Change jsdocs action to composite

* Remove composite step for jsdocs

* Remove docs directory

* Change branch ref to pipeline-phase-1.1

* Remove excluding docs directory from githook