Pipeline phase 1 (#34)
* Setup basic linting with ESLint and CodeClimate

This commit sets the recommended rules for ESLint and sets up the
CodeClimate configuration file. The default parser was changed to
Babel due to ESLint erroring with 'unexpected token' when using
the default parser provided. ESLint isn't configured yet to
reject/revise upon commit locally - the checks only run when
pushing to remote.

* Move codeclimate configuration file to root so it's detectable

This commit fixes the .codeclimate.yml file not being detectable
by CodeClimate.

* Added Jest, and make a couple example functions to test how Jest works, and to show some basic functionality of Jest

* testing actions

* added jest action

* Added JSDoc

* Setup ESLint extension for VSCode

This commit installs the necessary packages needed for the ESLint
extension so we can see style issues in the VSCode editor.
The changes that had to be made were:

 - Installing ESLint locally (npm install -g eslint)
 - Installing the babel parser locally
   (npm install eslint@4.x babel-eslint@8 --save-dev)
 - Adding the VSCode extension (all members will have to do this)
 - Adding a .eslintignore file to ignore certain directories

* Add script to run ESLint on all files before commit

This commit adds a git hook to run ESLint on all staged files
(modified, added, or copied) and report which checks failed and
which checks passed. It lets the user commit anyway if they choose
in case ESLint is being overly picky. The script is modified
from this link:

https://gist.github.com/linhmtran168/2286aeafe747e78f53bf

Git hooks are not tracked by git and are local in the .git/hooks
folder. To get around this I created a .githooks directory and
moved the pre-commit hook into that folder. Each person must
run the command git config --local core-hooksPath .githooks/
in order to change the default git hooks directory to this new
.githooks directory under version control. If anybody has another
way to get around this that would be more ideal but this is the
simplest solution I could find in the moment.

* Document how to set up linting

* Test ESLint github action

* Push to test ESLint github action

* Update phase1.md

* Update phase1.md

* Update phase1.md

* Update phase1.md

* Add files via upload

* Update phase1.md

* Update phase1.md

* Update phase1.md

* Add files via upload

* Update phase1.md

* Add ci pipeline diagram

* Upload ci pipeline video

* Merging all actions work into pipeline-phase-1 (#24)

This commit adds all the Github Action related changes regarding composite run step actions, and adds actions for linting, testing, and jsdocs.

* Pipeline phase 1 update deliverables (#25)

* Update diagram and status report (JSDocs)

* Add files via upload

Co-authored-by: zainnk902 <68146410+zainnk902@users.noreply.github.com>

Co-authored-by: Zain Khan <zainnk902@gmail.com>
Co-authored-by: dcullenUCSD <dcullen@ucsd.edu>
Co-authored-by: zainnk902 <68146410+zainnk902@users.noreply.github.com>
Co-authored-by: Dylan <60684055+dcullenUCSD@users.noreply.github.com>